Added a method to return CHIP8_VirtualMachine object to initial state upon repeated LoadProgram calls
Previously, the registers and main memory were not being set back to the default state when overwriting a *.ch8 program with a new one. This potential bug was not causing issues in CHIP-8 Player behaviour due to completely disposing of and recreating a CHIP8_VirtualMachine object when loading a new program.